# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Wheelhouse Tattoo Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inquiries:
    post:
      operationId: createInquiry
      tags: [inquiries]
      summary: Submit a customer inquiry
      description: Customer submits an event inquiry/booking request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInquiryBody"
      responses:
        "201":
          description: Inquiry submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
    get:
      operationId: listInquiries
      tags: [inquiries]
      summary: List all inquiries (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [new, contacted, quoted, booked, declined]
          description: Filter by status
        - name: state
          in: query
          schema:
            type: string
          description: Filter by state
        - name: packageType
          in: query
          schema:
            type: string
            enum: [A, B, C]
          description: Filter by package type
      responses:
        "200":
          description: List of inquiries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Inquiry"
  /inquiries/{id}:
    get:
      operationId: getInquiry
      tags: [inquiries]
      summary: Get inquiry details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inquiry details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateInquiryStatus
      tags: [inquiries]
      summary: Update inquiry status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInquiryStatusBody"
      responses:
        "200":
          description: Updated inquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - name: state
          in: query
          schema:
            type: string
          description: Filter by state
        - name: upcoming
          in: query
          schema:
            type: boolean
          description: Only return upcoming events
        - name: packageType
          in: query
          schema:
            type: string
            enum: [A, B, C]
          description: Filter by package type
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update an event (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /events/{id}/signup:
    post:
      operationId: signUpForEvent
      tags: [events]
      summary: Artist signs up for an event
      description: First come, first served — artist claims an open slot at the event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Signed up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistEventSignup"
        "409":
          description: Event full or already signed up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: withdrawFromEvent
      tags: [events]
      summary: Artist withdraws from an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Withdrawn
  /events/{id}/signups:
    get:
      operationId: getEventSignups
      tags: [events]
      summary: Get all artist signups for an event (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Artist signups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArtistEventSignup"
...
  /rigs:
    get:
      operationId: listRigs
      tags: [rigs]
      summary: List all rigs (admin)
      responses:
        "200":
          description: List of rigs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rig"
    post:
      operationId: createRig
      tags: [rigs]
      summary: Create a rig (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRigBody"
      responses:
        "201":
          description: Created rig
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rig"
  /rigs/{id}:
    put:
      operationId: updateRig
      tags: [rigs]
      summary: Update a rig (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRigBody"
      responses:
        "200":
          description: Updated rig
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rig"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        state:
          type: string
        city:
          type: string
        venue:
          type: ["string", "null"]
        packageType:
          type: string
          enum: [A, B, C]
        eventDate:
          type: string
          format: date-time
        artistSlots:
          type: integer
        signedUpCount:
          type: integer
        status:
          type: string
          enum: [open, full, completed, cancelled]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - state
        - city
        - packageType
        - eventDate
        - artistSlots
        - signedUpCount
        - status
        - createdAt
    CreateEventBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        state:
          type: string
        city:
          type: string
        venue:
          type: string
        packageType:
          type: string
          enum: [A, B, C]
        eventDate:
          type: string
          format: date-time
        artistSlots:
          type: integer
        notes:
          type: string
      required:
        - title
        - state
        - city
        - packageType
        - eventDate
        - artistSlots
    ArtistEventSignup:
      type: object
      properties:
        id:
          type: integer
        artistId:
          type: integer
        eventId:
          type: integer
        artistName:
          type: ["string", "null"]
        artistState:
          type: ["string", "null"]
        artistStyles:
          type: array
          items:
            type: string
        artistInstagram:
          type: ["string", "null"]
        eventTitle:
          type: ["string", "null"]
        eventDate:
          type: ["string", "null"]
          format: date-time
        eventState:
          type: ["string", "null"]
        eventCity:
          type: ["string", "null"]
        eventPackageType:
          type: ["string", "null"]
        signedUpAt:
          type: string
          format: date-time
      required:
        - id
        - artistId
        - eventId
        - artistStyles
        - signedUpAt
    Inquiry:
      type: object
      properties:
        id:
          type: integer
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: ["string", "null"]
        eventName:
          type: ["string", "null"]
        eventDate:
          type: ["string", "null"]
          format: date-time
        eventState:
          type: string
        eventCity:
          type: string
        expectedAttendees:
          type: ["integer", "null"]
        packageType:
          type: string
          enum: [A, B, C]
        message:
          type: ["string", "null"]
        status:
          type: string
          enum: [new, contacted, quoted, booked, declined]
        adminNotes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - contactName
        - contactEmail
        - eventState
        - eventCity
        - packageType
        - status
        - createdAt
    CreateInquiryBody:
      type: object
      properties:
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        eventName:
          type: string
        eventDate:
          type: string
          format: date-time
        eventState:
          type: string
        eventCity:
          type: string
        expectedAttendees:
          type: integer
        packageType:
          type: string
          enum: [A, B, C]
        message:
          type: string
      required:
        - contactName
        - contactEmail
        - eventState
        - eventCity
        - packageType
    UpdateInquiryStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [new, contacted, quoted, booked, declined]
        adminNotes:
          type: string
      required:
        - status
    Rig:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [available, deployed, maintenance]
        currentEventId:
          type: ["integer", "null"]
        currentEventTitle:
          type: ["string", "null"]
        homeState:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - status
        - homeState
        - createdAt
    CreateRigBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [available, deployed, maintenance]
        currentEventId:
          type: integer
        homeState:
          type: string
      required:
        - name
        - homeState
...