# lib/api-zod/src/generated/types/rig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wheelhouse Tattoo Platform API
 * OpenAPI spec version: 0.2.0
 */
import type { RigStatus } from "./rigStatus";

export interface Rig {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  status: RigStatus;
  /** @nullable */
  currentEventId?: number | null;
  /** @nullable */
  currentEventTitle?: string | null;
  homeState: string;
  createdAt: Date;
}